refactor: nightly maintainability and docs
Split the 1265-line lib/templates/templates.ts monolith into one file per
template (pov-life, sleep-story, finance-tips, true-crime,
pov-financial-lifestyle, kids-animated, psychology-of). templates.ts is now a
thin barrel that assembles the TEMPLATES array and re-exports getTemplateById
and the types. Shared Template/TemplateShowcase types and the templateAsset
helper moved to lib/templates/types.ts.

All existing exports and import paths are preserved unchanged. No behavior
changes.
