+++ lib/templates/kids-animated.ts
@@ -1,0 +1,12 @@
+import dedent from "dedent";
+import type { Template } from "./types";
+
+export const kidsAnimated: Template = {
+	id: "kids-animated",
+	name: "Kids Animated",
+	pillText: "A kids animated story about",
+	color: "#10B981",
+	referenceImages: [],
+	exampleText: "TODO",
+	systemPrompt: dedent`TODO`,
+};
