+++ lib/templates/psychology-of.ts
@@ -1,0 +1,12 @@
+import dedent from "dedent";
+import type { Template } from "./types";
+
+export const psychologyOf: Template = {
+	id: "psychology-of",
+	name: "Psychology",
+	pillText: "Psychology of",
+	color: "#EC4899",
+	referenceImages: [],
+	exampleText: "TODO",
+	systemPrompt: dedent`TODO`,
+};
