+++ lib/templates/true-crime.ts
@@ -1,0 +1,284 @@
+import dedent from "dedent";
+import type { Template } from "./types";
+
+export const trueCrime: Template = {
+	id: "true-crime",
+	name: "True Crime",
+	pillText: "A true crime story about",
+	color: "#8A0000",
+	style:
+		"Semi-realistic digital comic illustration, cel-shaded with bold ink outlines, muted earthy palette, cinematic dramatic lighting, gritty detailed textures, expressive characters, vertical 9:16 composition, Rockstar Games concept art style",
+	referenceImages: [
+		"https://mqzeech9ugknls54.public.blob.vercel-storage.com/assets/upload/template/true-crime-1",
+		"https://mqzeech9ugknls54.public.blob.vercel-storage.com/assets/upload/template/true-crime-2",
+		"https://mqzeech9ugknls54.public.blob.vercel-storage.com/assets/upload/template/true-crime-3",
+		"https://mqzeech9ugknls54.public.blob.vercel-storage.com/assets/upload/template/true-crime-4",
+	],
+	narration: {
+		gender: "masculine",
+		age: "adult",
+		pitch: "medium",
+		accent: "american",
+		description: "Friendly young adult male",
+	},
+	showcase: {
+		image:
+			"https://mqzeech9ugknls54.public.blob.vercel-storage.com/assets/upload/template/true-crime-5",
+		title: "True Crime",
+		description:
+			"Long-form, wild true crime stories told like a buddy spilling the craziest thing he ever heard",
+		examplePrompt: "Andres Escobar",
+	},
+	exampleText: dedent`
+#Music [tense, dramatic intro]
+
+#Image [a soccer ball lying in grass with blood splattered on it]
+#Narration: Imagine getting so mad about a soccer game that you kill someone.
+
+#Image [Andres Escobar smiling in a Colombia jersey, stadium behind him]
+#Narration: So the story starts with this guy, Andreas.
+
+#Image [Andres standing in front of a Colombian flag, looking friendly]
+#Narration: And Andreas is about 27, and he lives in Colombia.
+
+#Image [Andres in a soccer uniform kicking a ball on the field]
+#Narration: And he plays soccer for his country's big team.
+
+#Image [fans cheering and holding up Andres jerseys]
+#Narration: Everyone there loves him because he's a nice dude.
+
+#Image [Andres shaking hands politely with another player after a game]
+#Narration: He doesn't talk trash. He doesn't play dirty. He's just chill.
+
+#Image [a massive crowd of Colombian fans waving yellow flags in a city street]
+#Narration: And in Colombia, soccer is a HUGE deal. Like, a really huge deal.
+
+#Image [Andres on a magazine cover labeled "EL HÉROE"]
+#Narration: So Andreas isn't just famous. He's basically a hero.
+
+#Image [scoreboard showing Colombia 5, Argentina 0]
+#Narration: But then in 1993, his team smashes Argentina five to zero.
+
+#Image [Andres jumping in celebration with his teammates on the field]
+#Narration: Boom. That means Colombia gets to play in the World Cup.
+
+#Sound [crowd cheering, horns blaring]
+#Image [people partying in the streets of Bogotá, confetti everywhere]
+#Narration: And the whole country goes nuts.
+
+#Image [shadowy men in suits watching a small TV in a dim room]
+#Narration: But here's the problem. The bad guys are watching too.
+
+#Image [a chalkboard with betting numbers and stacks of cash on a table, angry cartel men around it]
+#Narration: The Colombian drug cartel, basically a giant crime gang, they bet millions of dollars on Colombia to win.
+
+#Image [Andres looking nervous, hand on his head in a locker room]
+#Narration: So now Andreas isn't just playing for fun. He's playing to keep some very scary people happy.
+
+#Music [music turns darker]
+
+#Image [airplane flying over California palm trees]
+#Narration: So in 1994, the team flies to California for the World Cup.
+
+#Image [scoreboard showing Romania 3, Colombia 1]
+#Narration: First game, Colombia versus Romania. And they lose. Three to one.
+
+#Image [Colombian fans crying in front of a TV at home]
+#Narration: The whole country is bummed.
+
+#Image [angry cartel boss slamming his fist on a desk, money flying]
+#Narration: The drug cartel? They are FURIOUS. They just lost a ton of money.
+
+#Image [a dark hotel hallway with one door slightly open]
+#Narration: And here's where things get really scary.
+
+#Image [hotel room TV glowing with a creepy threatening message on screen]
+#Narration: Back at the team's hotel, somebody hacks into the TVs.
+
+#Image [close-up of the TV screen showing skull and crossbones with a written warning]
+#Narration: And instead of a hello message, there's a threat. It says, "Don't let this one player play, or we'll kill all of you and bomb your families."
+
+#Sound [static, heartbeat]
+#Image [players sitting around the hotel room looking terrified]
+#Narration: The team is freaking out.
+
+#Image [Andres lacing up his cleats with a calm, focused face]
+#Narration: But Andreas, being the chill guy he is, he stays positive. He's gonna give it his all.
+
+#Image [huge packed stadium with Colombia and USA flags everywhere]
+#Narration: Next game. Colombia versus the USA. Ninety thousand fans in the stands.
+
+#Image [empty Colombian street, closed shop signs, only a glowing TV through a window]
+#Narration: Back home, stores are closed. Families are glued to the TV.
+
+#Image [cartel men watching the game intensely in a smoky room]
+#Narration: And yep, the cartel guys are watching too.
+
+#Music [intense game music]
+
+#Image [Andres sprinting on the soccer field, ball at his feet]
+#Narration: The game starts. Andreas is playing hard. Running, kicking, doing his thing.
+
+#Image [stadium clock showing 20 minutes, players mid-play]
+#Narration: But then, about 20 minutes in, something crazy happens.
+
+#Image [American player kicking a long pass across the field]
+#Narration: An American player tries to pass the ball.
+
+#Image [Andres sliding with his leg out toward the ball]
+#Narration: Andreas slides in to stop it.
+
+#Image [slow-motion close-up of ball bouncing off Andres's cleat at a weird angle]
+#Narration: But the ball bounces off his foot at a weird angle...
+
+#Sound [whoosh, then a net swish]
+#Image [the ball hitting the back of the Colombian net, goalie diving the wrong way]
+#Narration: ...and rolls right past his own goalie into his own net.
+
+#Image [Andres on his knees, hands on his head, devastated]
+#Narration: Yep. Andreas just scored on his own team.
+
+#Image [close-up of Andres's face, tears in his eyes]
+#Narration: And he knows right away. This is it. They're done.
+
+#Image [final scoreboard: USA 2, Colombia 1]
+#Narration: And so Colombia loses. And they get kicked out of the World Cup.
+
+#Image [angry Colombian fans burning jerseys in the street]
+#Narration: And the whole country is mad. And everybody starts pointing the finger at Andreas.
+
+#Sound [phone ringing creepily]
+#Image [a hotel phone ringing with a shadowy hand reaching for it]
+#Narration: Pretty quick, the players start getting scary phone calls.
+
+#Image [a player listening to the phone with a horrified face]
+#Narration: Like, "We're gonna hurt you" calls.
+
+#Image [the team eating dinner together at a US restaurant, looking nervous]
+#Narration: So the team decides to stay in America for a bit.
+
+#Image [Andres smiling and signing an autograph for a kid at an airport]
+#Narration: But Andreas? Nah, he's not scared. He's a positive dude.
+
+#Image [Andres walking off a plane in Colombia, suitcase in hand]
+#Narration: So he flies back to Colombia.
+
+#Image [Andres's mom hugging him in a kitchen, looking worried]
+#Narration: His friends, his family, even his coach beg him, "Please stay inside. It's not safe."
+
+#Image [Andres looking out a window with a determined expression]
+#Narration: But Andreas doesn't want to hide.
+
+#Image [a newspaper with Andres's column printed and his photo next to it]
+#Narration: He even writes a public letter saying sorry for the goal, and that "life doesn't end here."
+
+#Image [Andres looking up at the sky, a small sad smile on his face]
+#Narration: Welp. He had no idea how wrong he was about to be.
+
+#Music [darker, slower music]
+
+#Image [neon nightclub sign glowing in the dark]
+#Narration: Ten days later, Andreas goes out to a nightclub with his friends.
+
+#Image [Andres laughing with friends at a club table, drink in hand]
+#Narration: He's finally feeling okay again.
+
+#Image [a group of mean-looking men in fancy shirts glaring across the club]
+#Narration: But across the room, there's a group of bad guys, the Gallon brothers.
+
+#Image [the brothers counting stacks of cash earlier in a dim room]
+#Narration: They're in the cartel. And they bet a TON of money on that game. And they lost.
+
+#Image [the brothers pointing and yelling at Andres across the club]
+#Narration: They see Andreas and start yelling mean stuff at him, making fun of him for that goal.
+
+#Image [Andres calmly sipping his drink, looking the other way]
+#Narration: But Andreas, being chill, just ignores it.
+
+#Image [Andres standing up from his table, putting on his jacket]
+#Narration: Eventually he's like, "I'm out."
+
+#Image [Andres walking through a dark parking lot toward his car]
+#Narration: So he heads to the parking lot and gets in his car.
+
+#Image [the Gallon brothers and a huge bodyguard following him out the door]
+#Narration: But the bad guys follow him outside. They keep yelling.
+
+#Image [Andres holding his hands up, trying to explain himself]
+#Narration: And Andreas tries to explain, "Hey, it was just an accident. It could happen to anybody."
+
+#Sound [tense silence]
+#Image [the brothers' angry faces, fists clenched]
+#Narration: But they don't care. They lost millions. And they want somebody to blame.
+
+#Sound [blam blam blam blam blam blam]
+#Image [the bodyguard pulling a gun from his jacket, muzzle flash lighting up the lot]
+#Narration: And suddenly, their bodyguard pulls out a gun and shoots Andreas six times.
+
+#Image [the bodyguard yelling with a twisted, mocking face]
+#Narration: And after every single shot, he yells, "GOOOAL!" making fun of the mistake.
+
+#Image [crime scene tape stretched across a dark parking lot, car door open]
+#Narration: Then they jump in their car and drive off. And sadly, Andreas doesn't make it.
+
+#Image [a soccer ball sitting alone in an empty stadium]
+#Narration: All that... over a soccer game.
+
+#Image [bodyguard's mug shot under a police number board]
+#Narration: Now, the Gallon brothers ordered the hit, but their bodyguard takes the fall.
+
+#Image [the bodyguard in handcuffs being walked out by police]
+#Narration: And bam, he gets arrested. Here's his mug shot.
+
+#Image [the Gallon brothers smirking and sliding an envelope of cash to a man in a suit]
+#Narration: Now this part is gonna make you mad. The Gallon brothers? The guys who actually ordered the hit? They paid off the prosecutors.
+
+#Image [the brothers walking free outside a courthouse, smiling]
+#Narration: So they never got in trouble. Not one bit.
+
+#Image [the bodyguard walking out of prison with a duffel bag, only 11 years later]
+#Narration: And the bodyguard? He got sentenced to 43 years. But because of his connections, he only served 11.
+
+#Music [somber outro]
+#Image [a memorial statue of Andres Escobar with flowers piled at the base]
+#Narration: And that's the wild, sad story of how a soccer game ended a man's life.`,
+	systemPrompt: dedent`
+You write short narrative scripts in the style of viral YouTube true-story/crime videos. Pastiche these conventions precisely:
+
+# OPENING HOOK
+Open with a single punchy sentence that previews the wildest part of the story. Examples: "Imagine getting so mad over [X] that you murder someone." / "So this guy is about to [win/lose] [absurd thing] and then he's going to jail." / "So this man's obsession with [random thing] is about to go very wrong."
+
+# PROTAGONIST SETUP
+Immediately introduce the protagonist by first name, approximate age, and location: "Now, the guy's name is [Name]. And [Name] is about [X] when this story starts, and he's living in [Place]." Follow with a one-line problem statement: "And [Name] has a problem."
+
+# VOICE & TONE
+- Casual, conversational, like telling a buddy a wild story at a bar
+- First person narrator addressing the viewer directly ("I mean," "you know," "anyway")
+- Mild profanity
+- Sarcastic asides and dry humor ("Sure, Jan." / "I'm not making that up." / "give it a week or two")
+- Self-aware tangents and rhetorical questions ("And I don't know why this multi-millionaire doesn't have his own place, but whatever.")
+- Editorial reactions ("Damn." / "poor [Name]" / "this part's definitely going to make you mad")
+
+# PACING & STRUCTURE
+- Heavy use of "And," "So," "Now," "But then," "Anyway," and "And here's where things get [crazy/really out of control/scary]" as paragraph engines
+- Run-on sentences chained with "and" mixed with short punchy ones
+- Escalate events in clear beats, each worse or weirder than the last
+- Frequent reset phrases: "And so from there..." / "Here's where things get really out of control."
+- Maintain the rhythm: hook → setup → escalation → climax → fallout.
+
+# ONOMATOPOEIA
+Use written-out sound effects liberally and in clusters: blam blam blam, boom, bam, pow, kaboom, skirt, screech.
+
+# IMAGERY CUES
+The narrator should narrate the images that the viewer sees: "Here's his mug shot." / "Here's a picture of him." / "If you slow the body cam footage way down, you can see..." Use oddly specific numbers and dollar amounts for realism.
+
+# IMAGE MOTION
+Each static image should have a motion attribute, vary these as appropriate
+
+# IMAGE FREQUENCY
+Each narration sentence should have a different image associated with it
+
+# CLOSING
+Wrap up with the aftermath — arrest, trial, sentence, ironic twist, or grim ending — delivered matter-of-factly. Optional dry one-liner to button it ("All that because of a [thing].").
+`,
+};
